+++ src/serpent/convert/__init__.py
@@ -4,3 +4,4 @@
 from __future__ import annotations
 
 from .base64 import base64_to_num, num_to_base64
+from .digits import digits_to_num, num_to_digits
